Fix skl allocation to arena alignment
The code I copied from pebble had a const set for the alignment as
4 bytes. This worked fine since that implementation only used
uint32 atomics. However, this causes issues with this arch specific
implementation that uses either uint32 or uint64. To remedy this
I added `uint(unsafe.Sizeof(arch.UintToArchSize(0)))` to the node
alignment const. This will return a uint for whatever arch this
is compiled as giving us the proper byte alignment. 4 for 32-bit
and 8 for 64-bit.